packages/multi-tenant/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tplanet-ai-multi-tenant"
version = "0.1.0"
description = "Multi-tenant middleware and database router for Django"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "TPlanet Team", email = "dev@tplanet.ai" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["django", "multi-tenant", "saas", "middleware"]
dependencies = [
    "django>=4.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/town-intelligent/multi-tenant"
Documentation = "https://github.com/town-intelligent/multi-tenant#readme"
Repository = "https://github.com/town-intelligent/multi-tenant"

[tool.hatch.build.targets.wheel]
packages = ["django_multi_tenant"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
